exponent int instead of size_t

+++ python/rmm/_lib/memory_resource.pxd
@@ -57,8 +57,8 @@ cdef extern from "memory_resource_wrappers.hpp" nogil:
         ) except +
         binning_memory_resource_wrapper(
             shared_ptr[device_memory_resource_wrapper] upstream_mr,
-            size_t min_size_exponent,
-            size_t max_size_exponent
+            int min_size_exponent,
+            int max_size_exponent
         ) except +
         void add_bin(
             size_t allocation_size,
